der: refactor `tag` module; add tag number helpers
Performs some internal refactoring of the `tag` module, splitting out
`class` and `number` submodules.

Adds helper methods to `TagNumber` for creating specific classes of tags
with a given number.
